Various CI/packaging cleanups

+++ recipes/conda/build.sh
@@ -2,13 +2,15 @@
 
 mkdir build
 cd build
-cmake \
+cmake ${CMAKE_ARGS} \
     -DCMAKE_INSTALL_PREFIX=$PREFIX \
     -DCMAKE_INSTALL_LIBDIR=lib \
     -DPython_FIND_FRAMEWORK=NEVER \
     -DPython_FIND_STRATEGY=LOCATION \
     ..
 cmake --build .
+if [[ "${CONDA_BUILD_CROSS_COMPILATION}" != "1" ]]; then
 cmake --build . --target test
 rm -r /tmp/hdf5
+fi
 cmake --build . --target install
